// src/generated/api/artistObject.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Spotify Web API
 * You can use Spotify's Web API to discover music and podcasts, manage your Spotify library, control audio playback, and much more. Browse our available Web API endpoints using the sidebar at left, or via the navigation bar on top of this page on smaller screens.

In order to make successful Web API requests your app will need a valid access token. One can be obtained through <a href="https://developer.spotify.com/documentation/general/guides/authorization-guide/">OAuth 2.0</a>.

The base URI for all Web API requests is `https://api.spotify.com/v1`.

Need help? See our <a href="https://developer.spotify.com/documentation/web-api/guides/">Web API guides</a> for more information, or visit the <a href="https://community.spotify.com/t5/Spotify-for-Developers/bd-p/Spotify_Developer">Spotify for Developers community forum</a> to ask questions and connect with other developers.

 * OpenAPI spec version: 1.0.0
 */
import type { ExternalUrlObject } from './externalUrlObject';
import type { FollowersObject } from './followersObject';
import type { ImageObject } from './imageObject';
import type { ArtistObjectType } from './artistObjectType';

export interface ArtistObject {
  /** Known external URLs for this artist.
 */
  external_urls?: ExternalUrlObject;
  /** Information about the followers of the artist.
 */
  followers?: FollowersObject;
  /** A list of the genres the artist is associated with. If not yet classified, the array is empty.
 */
  genres?: string[];
  /** A link to the Web API endpoint providing full details of the artist.
 */
  href?: string;
  /** The [Spotify ID](/documentation/web-api/concepts/spotify-uris-ids) for the artist.
 */
  id?: string;
  /** Images of the artist in various sizes, widest first.
 */
  images?: ImageObject[];
  /** The name of the artist.
 */
  name?: string;
  /** The popularity of the artist. The value will be between 0 and 100, with 100 being the most popular. The artist's popularity is calculated from the popularity of all the artist's tracks.
 */
  popularity?: number;
  /** The object type.
 */
  type?: ArtistObjectType;
  /** The [Spotify URI](/documentation/web-api/concepts/spotify-uris-ids) for the artist.
 */
  uri?: string;
}