// src/generated/api/searchItemsResponse.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Spotify Web API
 * You can use Spotify's Web API to discover music and podcasts, manage your Spotify library, control audio playback, and much more. Browse our available Web API endpoints using the sidebar at left, or via the navigation bar on top of this page on smaller screens.

In order to make successful Web API requests your app will need a valid access token. One can be obtained through <a href="https://developer.spotify.com/documentation/general/guides/authorization-guide/">OAuth 2.0</a>.

The base URI for all Web API requests is `https://api.spotify.com/v1`.

Need help? See our <a href="https://developer.spotify.com/documentation/web-api/guides/">Web API guides</a> for more information, or visit the <a href="https://community.spotify.com/t5/Spotify-for-Developers/bd-p/Spotify_Developer">Spotify for Developers community forum</a> to ask questions and connect with other developers.

 * OpenAPI spec version: 1.0.0
 */
import type { PagingSimplifiedAlbumObject } from './pagingSimplifiedAlbumObject';
import type { PagingArtistObject } from './pagingArtistObject';
import type { PagingSimplifiedAudiobookObject } from './pagingSimplifiedAudiobookObject';
import type { PagingSimplifiedEpisodeObject } from './pagingSimplifiedEpisodeObject';
import type { PagingPlaylistObject } from './pagingPlaylistObject';
import type { PagingSimplifiedShowObject } from './pagingSimplifiedShowObject';
import type { PagingTrackObject } from './pagingTrackObject';

export type SearchItemsResponse = {
  albums?: PagingSimplifiedAlbumObject;
  artists?: PagingArtistObject;
  audiobooks?: PagingSimplifiedAudiobookObject;
  episodes?: PagingSimplifiedEpisodeObject;
  playlists?: PagingPlaylistObject;
  shows?: PagingSimplifiedShowObject;
  tracks?: PagingTrackObject;
};